Do not complain about missing network.
HTTP boot shall not complain about a missing link. The error message
overlaps the company logo. It is also misleading to call this "Error".
The user did not request to boot via HTTP and is not required to use the
network.  The PXE boot does not complain in a similar situation.